- removed warnings

+++ hi_modules/effects/editors/CurveEqEditor.h
@@ -62,7 +62,7 @@ public:
 		else
 		{
 			auto p = getProcessor()->getAttribute(parameter);
-			selectButton(p);
+			selectButton((int)p);
 		}
 	}
